requirements.txt
astroid==2.14.2
black==22.6.0
Bottleneck==1.3.5
brotlipy==0.7.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.0.4
click==8.0.4
colorama==0.4.6
cryptography==38.0.4
dill==0.3.6
filelock==3.9.0
flit_core==3.6.0
future==0.18.2
gmpy2==2.1.2
huggingface-hub==0.10.1
idna==3.4
isort==5.9.3
lazy-object-proxy==1.6.0
mccabe==0.7.0
mpmath==1.2.1
mypy==0.981
mypy-extensions==0.4.3
numexpr==2.8.4
numpy==1.23.5
packaging==22.0
pandas==1.5.2
pathspec==0.10.3
Pillow==9.3.0
pip==22.3.1
platformdirs==2.5.2
psutil==5.9.0
pycparser==2.21
pylint==2.16.2
pyOpenSSL==22.0.0
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2022.7
PyYAML==6.0
regex==2022.7.9
requests==2.28.1
setuptools==65.6.3
six==1.16.0
sympy==1.11.1
tokenizers==0.11.4
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.1
torch==2.0.0
torchaudio==2.0.1
torchvision==0.15.1
tqdm==4.64.1
transformers==4.24.0
typing_extensions==4.4.0
urllib3==1.26.14
wheel==0.38.4
wrapt==1.14.1
python-docx==0.8.11
bayoo-docx==0.2.20